Fix tests to ensure the .husk directory is always removed if it exists

+++ tests/cli.py
@@ -6,51 +6,53 @@ import unittest
 from husk import HuskError
 from husk.commands import init, add, remove, move, info
 
+def init_repo(*args):
+    options = init.parser.parse_args(args)
+    init.parser.handle_raw(options)
+    return options
 
-class CLIInitTestCase(unittest.TestCase):
+
+class BaseTestCase(unittest.TestCase):
     def tearDown(self):
-        shutil.rmtree('.husk')
+        if os.path.exists('.husk'):
+            shutil.rmtree('.husk')
 
+class CLIInitTestCase(BaseTestCase):
     def test_init_noargs(self):
-        options = init.parser.parse_args(''.split())
-        init.parser.handle_raw(options)
+        init_repo()
         self.assertTrue(os.path.exists('.husk'))
         self.assertFalse(os.path.exists('.husk/config'))
 
     def test_init_existing(self):
-        options = init.parser.parse_args(''.split())
-        init.parser.handle_raw(options)
+        options = init_repo()
         self.assertTrue(os.path.exists('.husk'))
         self.assertRaises(HuskError, init.parser.handle_raw, options)
 
     def test_init_path(self):
-        options = init.parser.parse_args(os.getcwd().split())
-        init.parser.handle_raw(options)
+        init_repo(os.getcwd())
         self.assertTrue(os.path.exists('.husk'))
 
     def test_init_config(self):
-        options = init.parser.parse_args('-d'.split())
-        init.parser.handle_raw(options)
+        init_repo('-d')
         self.assertTrue(os.path.exists('.husk/config'))
 
 
-class CLIAddTestCase(unittest.TestCase):
+class CLIAddTestCase(BaseTestCase):
     def test_add(self):
-        options = init.parser.parse_args()
-        init.parser.handle_raw(options)
+        init_repo()
 
         options = add.parser.parse_args('python/cli python/web'.split())
         add.parser.handle_raw(options)
+
         self.assertTrue(os.path.exists('python/cli/cues.md'))
         self.assertTrue(os.path.exists('python/web/summary.md'))
+
         shutil.rmtree('python')
-        shutil.rmtree('.husk')
 
 
-class CLIRemoveTestCase(unittest.TestCase):
+class CLIRemoveTestCase(BaseTestCase):
     def test_remove(self):
-        options = init.parser.parse_args()
-        init.parser.handle_raw(options)
+        init_repo()
 
         options = add.parser.parse_args('python/cli python/web'.split())
         add.parser.handle_raw(options)
@@ -63,13 +65,10 @@ class CLIRemoveTestCase(unittest.TestCase):
         remove.parser.handle_raw(options)
         self.assertFalse(os.path.exists('python/web'))
 
-        shutil.rmtree('.husk')
 
-
-class CLIMoveTestCase(unittest.TestCase):
+class CLIMoveTestCase(BaseTestCase):
     def test_move(self):
-        options = init.parser.parse_args()
-        init.parser.handle_raw(options)
+        init_repo()
 
         options = add.parser.parse_args('python/cli python/web'.split())
         add.parser.handle_raw(options)
@@ -79,14 +78,12 @@ class CLIMoveTestCase(unittest.TestCase):
         self.assertTrue(os.path.exists('python/command-line'))
         self.assertFalse(os.path.exists('python/cli'))
 
-        shutil.rmtree('.husk')
         shutil.rmtree('python')
 
 
-class CLIInfoTestCase(unittest.TestCase):
+class CLIInfoTestCase(BaseTestCase):
     def test_info(self):
-        options = init.parser.parse_args()
-        init.parser.handle_raw(options)
+        init_repo()
 
         options = add.parser.parse_args('python/cli python/web'.split())
         add.parser.handle_raw(options)
@@ -98,7 +95,6 @@ class CLIInfoTestCase(unittest.TestCase):
         self.assertEqual(sys.stdout.getvalue(), '2 notes\n-----\npython/cli\npython/web\n')
         sys.stdout = stdout
 
-        shutil.rmtree('.husk')
         shutil.rmtree('python')
